Fixed Memory Allocation error due to Rollover
There are two integers in Python: a 31 bit integer and an integer that
starts at 33 bits and could be larger.  The counter value cannot switch
between this two integer types because counter increment is called in an
interrupt, and a memory allocation error will occur if integer types
change.  So, best solution is to have the counter value roll over at the
largest possible positive small integer, which is 2**30-1 = 1073741823.
The user can put in a lower rollover value if desired.